Move ExperimentState and ExperimentStatus to run_model
