+++ src/ert/run_models/run_model.py
@@ -17,11 +17,12 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 from collections.abc import Callable, Generator, MutableSequence
 from contextlib import contextmanager
+from enum import StrEnum, auto
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, ClassVar, Protocol, cast
 
 import numpy as np
-from pydantic import PrivateAttr, field_validator
+from pydantic import BaseModel, PrivateAttr, field_validator
 from pydantic_core.core_schema import ValidationInfo
 
 from _ert.events import EEEvent, EESnapshot, EESnapshotUpdate
@@ -71,6 +72,20 @@ if TYPE_CHECKING:
     from ert.plugins import ErtRuntimePlugins
 
 logger = logging.getLogger(__name__)
+
+
+class ExperimentState(StrEnum):
+    pending = auto()
+    running = auto()
+    completed = auto()
+    stopped = auto()
+    failed = auto()
+    never_run = auto()
+
+
+class ExperimentStatus(BaseModel):
+    message: str = ""
+    status: ExperimentState = ExperimentState.pending
 
 
 class OutOfOrderSnapshotUpdateException(ValueError):
